Changelog more fixes
I have updated the header size and changed the boxes to be clickable, added the changelogs links, and got rid of the year range for a better year and quarter filters. I tried out the arrow in the bottom middle but it looks off as the right side is too empty, if yall wanna display any other info I can put it there and move the arrow.